Move the MacaroonGenerator
